+++ flathub.json
@@ -1,6 +1,5 @@
 {
     "only-arches": [
         "x86_64"
-    ],
-    "disable-external-data-checker": true
+    ]
 }
