+++ git-packages-lock.json
@@ -1,0 +1,42 @@
+[
+    {
+        "name": "bitcoin-safe-lib",
+        "archive_filename": "bitcoin_safe_lib-9f2636b19a44ca4ae01f8b0ff4c0582fe5cb93ad.tar.gz",
+        "url": "https://github.com/andreasgriffin/bitcoin-safe-lib/archive/9f2636b19a44ca4ae01f8b0ff4c0582fe5cb93ad.tar.gz"
+    },
+    {
+        "name": "hwi",
+        "archive_filename": "hwi-38ccb5169365d2a635539ef828ab46ed81dd1205.tar.gz",
+        "url": "https://github.com/andreasgriffin/HWI/archive/38ccb5169365d2a635539ef828ab46ed81dd1205.tar.gz"
+    },
+    {
+        "name": "jade_client",
+        "archive_filename": "jade_client-fb2642397ed7afdda08f3d1380ac6ee21164e7e6.tar.gz",
+        "url": "https://github.com/Blockstream/Jade/archive/fb2642397ed7afdda08f3d1380ac6ee21164e7e6.tar.gz"
+    },
+    {
+        "name": "bitcoin-usb",
+        "archive_filename": "bitcoin_usb-26450e4fa299dcbb0073360558c67a7ef17d8951.tar.gz",
+        "url": "https://github.com/andreasgriffin/bitcoin-usb/archive/26450e4fa299dcbb0073360558c67a7ef17d8951.tar.gz"
+    },
+    {
+        "name": "pyzbar",
+        "archive_filename": "pyzbar-bad840c31dd0f92b616ca29642a408804d5cfa37.tar.gz",
+        "url": "https://github.com/andreasgriffin/pyzbar/archive/bad840c31dd0f92b616ca29642a408804d5cfa37.tar.gz"
+    },
+    {
+        "name": "bitcoin-qr-tools",
+        "archive_filename": "bitcoin_qr_tools-a3d10dff6d7257240f1bfded36ed8aad6be172da.tar.gz",
+        "url": "https://github.com/andreasgriffin/bitcoin-qr-tools/archive/a3d10dff6d7257240f1bfded36ed8aad6be172da.tar.gz"
+    },
+    {
+        "name": "bitcoin-nostr-chat",
+        "archive_filename": "bitcoin_nostr_chat-ae05c7ee97371415b4bf5e9aacf4a2c7ed86eb1d.tar.gz",
+        "url": "https://github.com/andreasgriffin/bitcoin-nostr-chat/archive/ae05c7ee97371415b4bf5e9aacf4a2c7ed86eb1d.tar.gz"
+    },
+    {
+        "name": "btcpay-tools",
+        "archive_filename": "btcpay_tools-ad910a68616288358dea1f614916d92addcefbf3.tar.gz",
+        "url": "https://github.com/andreasgriffin/btcpay-tools/archive/ad910a68616288358dea1f614916d92addcefbf3.tar.gz"
+    }
+]
